// front-end/src/api/models/api/dailyTodoOption.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Wijckie API
 * Default API for Wijckie
 * OpenAPI spec version: 1.0.0
 */

export interface DailyTodoOption {
    readonly id: number
    readonly module: number
    /**
     * @minLength 1
     * @maxLength 30
     */
    name: string
    text: string
}